+++ apps/mobile/app/common/database/index.js
@@ -49,16 +49,16 @@ export var db = new Database(
 db.host(
   __DEV__
     ? {
-        // API_HOST: "https://api.notesnook.com",
-        // AUTH_HOST: "https://auth.streetwriters.co",
-        // SSE_HOST: "https://events.streetwriters.co",
-        // SUBSCRIPTIONS_HOST: "https://subscriptions.streetwriters.co",
-        // ISSUES_HOST: "https://issues.streetwriters.co"
-        API_HOST: "http://192.168.8.109:5264",
-        AUTH_HOST: "http://192.168.8.109:8264",
-        SSE_HOST: "http://192.168.8.109:7264",
-        SUBSCRIPTIONS_HOST: "http://192.168.8.109:9264",
-        ISSUES_HOST: "http://192.168.8.109:2624"
+        API_HOST: "https://api.notesnook.com",
+        AUTH_HOST: "https://auth.streetwriters.co",
+        SSE_HOST: "https://events.streetwriters.co",
+        SUBSCRIPTIONS_HOST: "https://subscriptions.streetwriters.co",
+        ISSUES_HOST: "https://issues.streetwriters.co"
+        // API_HOST: "http://192.168.8.109:5264",
+        // AUTH_HOST: "http://192.168.8.109:8264",
+        // SSE_HOST: "http://192.168.8.109:7264",
+        // SUBSCRIPTIONS_HOST: "http://192.168.8.109:9264",
+        // ISSUES_HOST: "http://192.168.8.109:2624"
       }
     : {
         API_HOST: "https://api.notesnook.com",
