+++ apps/mobile/app/components/auth/session-expired.js
@@ -17,37 +17,32 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { Modal, View } from "react-native";
 import { db } from "../../common/database";
 import BiometricService from "../../services/biometrics";
 import {
-  eSendEvent,
   eSubscribeEvent,
   eUnSubscribeEvent,
   ToastEvent
 } from "../../services/event-manager";
-import { clearMessage } from "../../services/message";
-import PremiumService from "../../services/premium";
 import SettingsService from "../../services/settings";
 import Sync from "../../services/sync";
 import { useThemeStore } from "../../stores/use-theme-store";
-import { useUserStore } from "../../stores/use-user-store";
 import { SIZE } from "../../utils/size";
 import { sleep } from "../../utils/time";
 import { Dialog } from "../dialog";
 import { presentDialog } from "../dialog/functions";
 import SheetProvider from "../sheet-provider";
-import { Progress } from "../sheets/progress";
 import { Toast } from "../toast";
 import { Button } from "../ui/button";
 import { IconButton } from "../ui/icon-button";
 import Input from "../ui/input";
 import Heading from "../ui/typography/heading";
 import Paragraph from "../ui/typography/paragraph";
-import TwoFactorVerification from "./two-factor";
+import { LoginSteps, useLogin } from "./use-login";
 
-function getEmail(email) {
+function getObfuscatedEmail(email) {
   if (!email) return null;
   return email.replace(/(.{2})(.*)(?=@)/, function (gp1, gp2, gp3) {
     for (let i = 0; i < gp3.length; i++) {
@@ -59,15 +54,14 @@ function getEmail(email) {
 
 export const SessionExpired = () => {
   const colors = useThemeStore((state) => state.colors);
-  const email = useRef();
-  const passwordInputRef = useRef();
-  const password = useRef();
   const [visible, setVisible] = useState(false);
   const [focused, setFocused] = useState(false);
-
-  const setUser = useUserStore((state) => state.setUser);
-
-  const [loading, setLoading] = useState(false);
+  const { step, password, email, passwordInputRef, loading, login } = useLogin(
+    undefined,
+    () => {
+      setVisible(false);
+    }
+  );
 
   const logout = async () => {
     try {
@@ -92,9 +86,9 @@ export const SessionExpired = () => {
       eUnSubscribeEvent("session_expired", open);
       setFocused(false);
     };
-  }, [visible]);
+  }, [visible, open]);
 
-  const open = async () => {
+  const open = React.useCallback(async () => {
     try {
       let res = await db.user.tokenManager.getToken();
       if (!res) throw new Error("no token found");
@@ -120,64 +114,8 @@ export const SessionExpired = () => {
       email.current = user.email;
       setVisible(true);
     }
-  };
+  }, [email]);
 
-  const login = async (mfa, callback) => {
-    setLoading(true);
-    let user;
-    try {
-      if (mfa) {
-        await db.user.mfaLogin(
-          email.current.toLowerCase(),
-          password.current,
-          mfa
-        );
-      } else {
-        await db.user.login(email.current.toLowerCase(), password.current);
-      }
-      callback && callback(true);
-      setVisible(false);
-      user = await db.user.getUser();
-      if (!user) throw new Error("Email or password incorrect!");
-      PremiumService.setPremiumStatus();
-      setUser(user);
-      clearMessage();
-      ToastEvent.show({
-        heading: "Login successful",
-        message: `Logged in as ${user.email}`,
-        type: "success",
-        context: "global"
-      });
-      await SettingsService.set({
-        sessionExpired: false,
-        userEmailConfirmed: user?.isEmailConfirmed
-      });
-      eSendEvent("userLoggedIn", true);
-      await sleep(500);
-      Progress.present();
-      setLoading(false);
-    } catch (e) {
-      callback && callback(false);
-      if (e.message === "Multifactor authentication required.") {
-        TwoFactorVerification.present(async (mfa) => {
-          if (mfa) {
-            await login(mfa);
-          } else {
-            setLoading(false);
-          }
-        }, e.data);
-      } else {
-        console.log(e.stack);
-        setLoading(false);
-        ToastEvent.show({
-          heading: user ? "Failed to sync" : "Login failed",
-          message: e.message,
-          type: "error",
-          context: "local"
-        });
-      }
-    }
-  };
   return (
     visible && (
       <Modal
@@ -194,7 +132,8 @@ export const SessionExpired = () => {
             width: focused ? "100%" : "99.9%",
             padding: 12,
             justifyContent: "center",
-            flex: 1
+            flex: 1,
+            backgroundColor: colors.bg
           }}
         >
           <View
@@ -225,29 +164,32 @@ export const SessionExpired = () => {
               }}
             >
               Your session on this device has expired. Please enter password for{" "}
-              {getEmail(email.current)} to continue.
+              {getObfuscatedEmail(email.current)} to continue.
             </Paragraph>
           </View>
 
-          <Input
-            fwdRef={passwordInputRef}
-            onChangeText={(value) => {
-              password.current = value;
-            }}
-            returnKeyLabel="Next"
-            returnKeyType="next"
-            secureTextEntry
-            autoComplete="password"
-            autoCapitalize="none"
-            autoCorrect={false}
-            placeholder="Password"
-            onSubmit={() => login()}
-          />
+          {step === LoginSteps.passwordAuth ? (
+            <Input
+              fwdRef={passwordInputRef}
+              onChangeText={(value) => {
+                password.current = value;
+              }}
+              returnKeyLabel="Next"
+              returnKeyType="next"
+              secureTextEntry
+              autoComplete="password"
+              autoCapitalize="none"
+              autoCorrect={false}
+              placeholder="Password"
+              onSubmit={() => login()}
+            />
+          ) : null}
 
           <Button
             style={{
               marginTop: 10,
-              width: "100%"
+              width: 250,
+              borderRadius: 100
             }}
             loading={loading}
             onPress={() => login()}
